defunct(ansible/aws): attempt to set provider to external
All of the docs for how to on-board the AWS provider say to set the
cloud-provider to external, however, I couldn't get this to work. Doing
so for the node would have been really complicated because I would have
to parse the join command to make a full JoinNode configuration document
for kubeadm. And doing it just on the control-node means that the worker
isn't able to join for some reason. It just sits there trying to join,
but says that the token doesn't work. Additionally, trying to get both
the config and the flags to work at the same time doesn't appear to
work.

At the end of the day, I primarily needed this just to get kube-router
to execute it's logic to remove the source-dest check. Now that I
control that via terraform, this isn't really all that important for me
any longer. Additionally, the AWS cloud provider still seems to work. I
have yet to test the cloud provider with LoadBalancer IPs, but it seems
like it should work, there are no errors in the logs.