fix(ansible): update to containerd.io
Use containerd from the official docker apt repo to get a more
up-to-date version than what is carried in Ubuntu. After upgrading to
Ubuntu 22.04 I wasn't able to get the official version of containerd to
even work any longer with kubeadm. Additionally, ensure that containerd
is not just started, but restarted as containerd.io appears to start it
automatically or in some way that precludes it from reading the config
we write before we get to the ansible command that starts it.